+++ Interface/IScript.h
@@ -978,93 +978,93 @@ namespace PaintsNow {
 
 #if defined(_MSC_VER) && _MSC_VER <= 1200
 			template <class A = Void, class B = Void, class C = Void, class D = Void, class E = Void, class F = Void, class G = Void, class H = Void, class I = Void, class J = Void, class K = Void, class L = Void, class M = Void>
-			struct Dispatcher : public TWrapper<std::pair<Request::Ref, uint32_t>, Request&, bool> {
+			struct Dispatcher : public TWrapper<std::pair<Request::Ref, size_t>, Request&, bool> {
 				void Invoke0(const Request::AutoWrapperBase& wrapper, Request& request) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second);
 					(*this)(request, false);
 				}
 
 				void Invoke1(const Request::AutoWrapperBase& wrapper, Request& request, A a) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a);
 					(*this)(request, false);
 				}
 
 				void Invoke2(const Request::AutoWrapperBase& wrapper, Request& request, A a, B b) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a, b);
 					(*this)(request, false);
 				}
 
 				void Invoke3(const Request::AutoWrapperBase& wrapper, Request& request, A a, B b, C c) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a, b, c);
 					(*this)(request, false);
 				}
 
 				void Invoke4(const Request::AutoWrapperBase& wrapper, Request& request, A a, B b, C c, D d) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a, b, c, d);
 					(*this)(request, false);
 				}
 
 				void Invoke5(const Request::AutoWrapperBase& wrapper, Request& request, A a, B b, C c, D d, E e) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a, b, c, d, e);
 					(*this)(request, false);
 				}
 
 				void Invoke6(const Request::AutoWrapperBase& wrapper, Request& request, A a, B b, C c, D d, E e, F f) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a, b, c, d, e, f);
 					(*this)(request, false);
 				}
 
 				void Invoke7(const Request::AutoWrapperBase& wrapper, Request& request, A a, B b, C c, D d, E e, F f, G g) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a, b, c, d, e, f, g);
 					(*this)(request, false);
 				}
 
 				void Invoke8(const Request::AutoWrapperBase& wrapper, Request& request, A a, B b, C c, D d, E e, F f, G g, H h) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a, b, c, d, e, f, g, h);
 					(*this)(request, false);
 				}
 
 				void Invoke9(const Request::AutoWrapperBase& wrapper, Request& request, A a, B b, C c, D d, E e, F f, G g, H h, I i) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a, b, c, d, e, f, g, h, i);
 					(*this)(request, false);
 				}
 
 				void Invoke10(const Request::AutoWrapperBase& wrapper, Request& request, A a, B b, C c, D d, E e, F f, G g, H h, I i, J j) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a, b, c, d, e, f, g, h, i, j);
 					(*this)(request, false);
 				}
 
 				void Invoke11(const Request::AutoWrapperBase& wrapper, Request& request, A a, B b, C c, D d, E e, F f, G g, H h, I i, J j, K k) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a, b, c, d, e, f, g, h, i, j, k);
 					(*this)(request, false);
 				}
 
 				void Invoke12(const Request::AutoWrapperBase& wrapper, Request& request, A a, B b, C c, D d, E e, F f, G g, H h, I i, J j, K k, L l) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a, b, c, d, e, f, g, h, i, j, k, l);
 					(*this)(request, false);
 				}
 
 				void Invoke13(const Request::AutoWrapperBase& wrapper, Request& request, A a, B b, C c, D d, E e, F f, G g, H h, I i, J j, K k, L l, M m) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a, b, c, d, e, f, g, h, i, j, k, l);
 					(*this)(request, false);
 				}
 
 				void Invoke14(const Request::AutoWrapperBase& wrapper, Request& request, A a, B b, C c, D d, E e, F f, G g, H h, I i, J j, K k, L l, M m) {
-					std::pair<Request::Ref, uint32_t> ref = (*this)(request, true);
+					std::pair<Request::Ref, size_t> ref = (*this)(request, true);
 					request.Call(wrapper, ref.first, ref.second, a, b, c, d, e, f, g, h, i, j, k, l, m);
 					(*this)(request, false);
 				}
@@ -1126,9 +1126,9 @@ namespace PaintsNow {
 			}
 #else
 			template <typename... Args>
-			struct Dispatcher : public TWrapper<std::pair<Request::Ref, uint32_t>, Request&, bool> {
+			struct Dispatcher : public TWrapper<std::pair<Request::Ref, size_t>, Request&, bool> {
 				void Invoke(const Request::AutoWrapperBase& wrapper, Request& request, Args&&... args) {
-					std::pair<IScript::Request::Ref, uint32_t> r = (*this)(request, true);
+					std::pair<IScript::Request::Ref, size_t> r = (*this)(request, true);
 					request.Call(wrapper, r.first, r.second, std::forward<Args>(args)...);
 					(*this)(request, false);
 				}
@@ -1246,7 +1246,7 @@ namespace PaintsNow {
 
 		class RequestPool {
 		public:
-			RequestPool(IScript& script, uint32_t size);
+			RequestPool(IScript& script, size_t size);
 			~RequestPool();
 			Request* AllocateRequest();
 			void FreeRequest(Request* request);
@@ -1256,8 +1256,8 @@ namespace PaintsNow {
 		protected:
 			IScript& script;
 			std::stack<Request*> requests;
-			uint32_t size;
-			std::atomic<int32_t> requestCritical;
+			size_t size;
+			std::atomic<size_t> requestCritical;
 		};
 
 	protected:
